% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpsTemplates.R
\docType{package}
\name{mpsTemplates-package}
\alias{mpsTemplates}
\alias{mpsTemplates-package}
\title{mpsTemplates: Templates for Me}
\description{
The mpsTemplates package provides opinionated ggplot themes
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mps9506.github.io/mps-templates/}
}

}
\author{
\strong{Maintainer}: Michael Schramm \email{mpschramm@gmail.com}

}
