% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mps_noto_dark.R
\name{noto_dark_geom_defaults}
\alias{noto_dark_geom_defaults}
\title{Change geom defaults from black to custom lights}
\usage{
noto_dark_geom_defaults()
}
\description{
Change geom defaults from black to custom lights
}
