% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mps_noto.R
\name{theme_mps_noto}
\alias{theme_mps_noto}
\title{an opinionated \link{ggplot2} theme}
\usage{
theme_mps_noto(
  base_size = 11,
  base_family = "NotoSansDisplay_Condensed",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_size}{default text size in px}

\item{base_family}{defaults to \code{"NotoSansDisplay_Condensed"}. Which is
provided in the package. Generally, use a sensible sans serif font that is
available on your system.}

\item{base_line_size}{default line size}

\item{base_rect_size}{default rect size}
}
\description{
Provides a \link{ggplot2} theme.
}
\details{
Uses Google's Noto Sans Display font by default. Using the
\code{\link[ragg:agg_png]{ragg::agg_png()}} or related graphics device should ensure the font is
properly detected and used.
}
\examples{
library(ggplot2)
library(ragg)

file <- tempfile(fileext = '.png')
agg_png(file)
ggplot(cars) +
  geom_point(aes(speed, dist)) +
  theme_mps_noto() +
  labs(title = "Cars", subtitle = "Speed and stopping distance",
       caption = "Ezekiel, M (1930) Methods of Correlation Analysis",
       x = "Speed [mph]",
       y = "Stopping Distance [ft]")
dev.off()
}
