% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mps_noto.R
\name{theme_mps_noto}
\alias{theme_mps_noto}
\title{an opinionated [ggplot2] theme}
\usage{
theme_mps_noto(
  base_size = 11,
  base_family = "NotoSansDisplay_Condensed",
  base_line_size = 0.5,
  base_rect_size = 0.5
)
}
\arguments{
\item{base_size}{default text size in px}

\item{base_family}{defaults to \code{"NotoSansDisplay_Condensed"}. Which is provided in the package. Generally, use a sensible sans serif font that is available on your system.}

\item{base_line_size}{default line size}

\item{base_rect_size}{default rect size}
}
\description{
Provides a [ggplot2] theme.
}
\examples{
ggplot(cars) +
  geom_point(aes(speed, dist)) +
  theme_mps_noto()
}
